correct full tests

+++ src/main/java/com/example/auto_concierge/service/ServiceCenterService.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -31,26 +32,41 @@ public class ServiceCenterService {
 
     public ServiceCenter createServiceCenter(Long userId, ServiceCenter serviceCenter) {
         User user = userRepository.findById(userId).orElse(null);
-        if (user != null && user.getRole() == Role.SERVICE_CENTER) {
+        if (user == null) {
+            throw new RuntimeException("Пользователь с идентификатором " + userId + " не найден");
+        }
+        if (user.getRole() == Role.SERVICE_CENTER) {
             if (serviceCenterRepository.existsByAddress(serviceCenter.getAddress())) {
                 throw new RuntimeException("СТО с таким адресом уже существует");
             }
             serviceCenter.setOwner(user);
             return serviceCenterRepository.save(serviceCenter);
         } else {
-            throw new RuntimeException("Недостаточно прав для создания сервис центра или пользователь с указанным ID не найден");
+            throw new RuntimeException("Недостаточно прав для создания сервис центра");
         }
     }
 
     public List<ServiceCenter> getAllServiceCenters() {
-        return serviceCenterRepository.findAll();
+        List<ServiceCenter> serviceCenters = serviceCenterRepository.findAll();
+        if (serviceCenters.isEmpty()) {
+            throw new RuntimeException("Сервисные центры не найдены");
+        }
+        return serviceCenters;
     }
 
-    public ServiceCenter getServiceCenterById(Long id) { return serviceCenterRepository.findById(id).orElse(null);}
+    public ServiceCenter getServiceCenterById(Long id) {
+        Optional<ServiceCenter> optionalServiceCenter = serviceCenterRepository.findById(id);
+        if (optionalServiceCenter.isEmpty()) {
+            throw new IllegalArgumentException("Сервисный центр с идентификатором " + id + " не найден");
+        }
+        return serviceCenterRepository.findById(id).orElse(null);
+    }
 
     public ServiceCenter updateServiceCenter(Long id, ServiceCenter serviceCenterDetails) {
         ServiceCenter serviceCenter = serviceCenterRepository.findById(id).orElse(null);
-        if (serviceCenter != null) {
+        if (serviceCenter == null) {
+            throw new IllegalArgumentException("Сервисный центр с идентификатором " + id + " не найден");
+        }
             serviceCenter.setName(serviceCenterDetails.getName());
             serviceCenter.setAddress(serviceCenterDetails.getAddress());
             serviceCenter.setContactNumber(serviceCenterDetails.getContactNumber());
@@ -58,13 +74,17 @@ public class ServiceCenterService {
             serviceCenter.setAverageRating(serviceCenterDetails.getAverageRating());
             serviceCenter.setSchedule(serviceCenterDetails.getSchedule());
             return serviceCenterRepository.save(serviceCenter);
-        } else {
-            throw new RuntimeException("Service center not found");
-        }
     }
 
     public void deleteServiceCenter(Long id) {
-        serviceCenterRepository.deleteById(id);
+        if (!serviceCenterRepository.existsById(id)) {
+            throw new IllegalArgumentException("Сервисный центр с идентификатором " + id + " не найден");
+        }
+        try {
+            serviceCenterRepository.deleteById(id);
+        } catch (Exception e) {
+            throw new RuntimeException("Ошибка при удалении сервисного центра с идентификатором " + id + ": " + e.getMessage(), e);
+        }
     }
 
     public List<ServiceCenter> findServiceCentersWithinRadius(double latitude, double longitude, double radius) {
@@ -72,26 +92,28 @@ public class ServiceCenterService {
         List<ServiceCenter> serviceCentersWithinRadius = new ArrayList<>();
         double radiusInKm = radius * unknownKoefficient;
         for (ServiceCenter serviceCenter : serviceCenters) {
-            double serviceCenterLat = Math.toRadians(serviceCenter.getAddress().getLatitude());
-            double serviceCenterLon = Math.toRadians(serviceCenter.getAddress().getLongitude());
-            double deltaLat = Math.toRadians(latitude - serviceCenterLat);
-            double deltaLon = Math.toRadians(longitude - serviceCenterLon);
-
-            double a = Math.sin(deltaLat / 2) * Math.sin(deltaLat / 2) +
-                    Math.cos(serviceCenterLat) * Math.cos(latitude) *
-                            Math.sin(deltaLon / 2) * Math.sin(deltaLon / 2);
-
-            double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-
-            double distance = earthRadius * c;
-
-            // Если расстояние меньше или равно заданному радиусу, добавляем сервис-центр в список
+            double distance = getDistance(latitude, longitude, serviceCenter);
             if (distance <= radiusInKm) {
                 serviceCentersWithinRadius.add(serviceCenter);
             }
         }
-
         return serviceCentersWithinRadius;
+    }
+
+    private static double getDistance(double latitude, double longitude, ServiceCenter serviceCenter) {
+        double serviceCenterLat = Math.toRadians(serviceCenter.getAddress().getLatitude());
+        double serviceCenterLon = Math.toRadians(serviceCenter.getAddress().getLongitude());
+        double deltaLat = Math.toRadians(latitude - serviceCenterLat);
+        double deltaLon = Math.toRadians(longitude - serviceCenterLon);
+
+        double a = Math.sin(deltaLat / 2) * Math.sin(deltaLat / 2) +
+                Math.cos(serviceCenterLat) * Math.cos(latitude) *
+                        Math.sin(deltaLon / 2) * Math.sin(deltaLon / 2);
+
+        double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+
+        double distance = earthRadius * c;
+        return distance;
     }
 }
